Refactor test_moe_qat_parallel
